pyproject.toml
[project]
name = "comfyui-gemini-edit"
version = "1.0.0"
description = "ComfyUI custom node for editing images using Google Gemini 2.5 Flash Image API"
authors = [
    { name = "ComfyUI Gemini Integration" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Pillow>=9.0.0",
    "numpy>=1.19.0",
    "requests>=2.25.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"